fix: 清理 Rust clippy 警告 41→8, 删真死代码, 标注未接线功能
自动修复(cargo clippy --fix): Ok(...?) 冗余包装/unused import/loop→for/
闭包复制/多余类型转换等 21 处机械项。

手工修复(4处, 均有实测依据):
- domain/bookpack.rs: 删 BookpackMeta 结构体(全仓库零引用, check_version 从
  未使用它, 直接操作 serde_json::Value)
- application/wizard_service.rs: 删 WIZARD_ID 常量(全仓库零引用, 连 SQL 里
  都是字面量 'first-run')
- jobs/spawn.rs: filter_map(|l| l.ok()) → map_while(Result::ok), 修复潜在的
  "持续读错误时死循环不终止"问题(clippy lines_filter_map_ok 指出)
- services/sync.rs: 删孤儿文档注释(描述"合并本地+远端"逻辑, 但紧跟的函数是
  urlencode; 文件头已声明该逻辑权威实现在 domain/sync.rs, 此注释纯冗余)

发现并标注 7 处"已写测试但从未被 command/前端调用"的函数(#[allow(dead_code)]
+ TODO 注释, 不删不装作没看见): bundle_complete/book_bundle_complete(模型包
完整性)、is_done(向导完成态)、delete_cf_token(CF断开连接)、
github_release_asset_url/hf_resolve_url(首次下载URL构造)、ProfileRepo::get。
前4类已登记为独立任务, 待用户决定是否/何时接线。

剩余 8 处警告(参数过多/类型复杂)全部集中在 jobs.rs 等命令层, 是已知的
"业务编排住在命令层"架构问题, 留给后续拆分 jobs.rs 时一并处理, 不在本次
lint 清理中勉强拆分。

回归验证: cargo fmt --check 干净; cargo test --release --test-threads=1
100 全绿(前后一致)。

+++ src-tauri/src/store/vocab_repo.rs
@@ -144,7 +144,7 @@ impl<'a> VocabRepo<'a> {
                 norm.word, norm.lemma, norm.pos, norm.meaning, norm.sense_id,
                 norm.phonetic, norm.context, norm.level,
                 serde_json::to_string(&norm.collocations).unwrap_or_else(|_| "[]".to_string()),
-                norm.stage, norm.interval as f64, norm.ease_factor, norm.next_review, norm.reviews,
+                norm.stage, norm.interval, norm.ease_factor, norm.next_review, norm.reviews,
                 norm.added_at, now, profile_id,
                 serde_json::to_string(&payload).unwrap_or_else(|_| "{}".to_string()),
             ],
